projects/daily-note-generator/requirements.txt
# Core dependencies
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.2
click>=8.1.7
rich>=13.7.0

# Date/time handling
python-dateutil>=2.8.2

# Optional enhancements
pytz>=2023.3

# Development dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
